advanced_rag_llama_index/requirements.txt
openai
python-dotenv
llama-index
trulens-eval